made changes to be able to merge with PR #20 .
